requirements.txt
youtube-transcript-api==1.2.2
python-dotenv==1.0.1